Worker calculates average CC for a given commit
Worker asks Master for a commit SHA. It downloads the file tree for the commit. For each Blob ('file') in the tree, separate the python and non python files. Get the file contents from Github for each python file and add it to a list. Create temp files and run Radon CC on each file and compute average. Clean up tmp dir afterwards